% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidviz-package.R
\docType{package}
\name{covidviz-package}
\alias{covidviz}
\alias{covidviz-package}
\title{covidviz: COVID Data Visualization}
\description{
It has a few functions that perform some non-trivial data wrangling and produce plots as results.
}
\author{
\strong{Maintainer}: Tom Zhang \email{zhangsiyuan777@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
