% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query1}
\alias{query1}
\title{Query 1}
\usage{
query1(
  plot_type = "static",
  fill = "cases",
  log_scale = FALSE,
  projection = "mercator",
  palette = "YlOrBr"
)
}
\arguments{
\item{plot_type}{"static" | "dynamic"}

\item{fill}{"cases" |"deaths" |"days" - The map will be filled by this variable.}

\item{log_scale}{FALSE | TRUE - Apply a log transformation on the selected variable? Could be useful if variable contains outliers.}

\item{projection}{"mercator" | "globular" | "gilbert"}

\item{palette}{"YlOrBr" or any sequential palette from RColorBrewer: "YlOrRd", "YlOrBr", "YlGnBu", "YlGn", "Reds", "RdPu", "Purples", "PuRd", "PuBuGn", "PuBu", "OrRd", "Oranges", "Greys", "Greens", "GnBu", "BuPu", "BuGn", "Blues".}
}
\value{
Either a static map produced with ggplot2 or a dynamic one produced with leaflet.
}
\description{
For each country, count the total number of cases, deaths, and the number of days passed since the first day of the outbreak (day of the first recorded new case) up to the day of the first peak stringency index (both inclusive).
}
