% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query2}
\alias{query2}
\title{Query 2}
\usage{
query2(
  plot_type = "static",
  fill = "avg_daily_new_cases_pre_vac",
  log_scale = FALSE,
  projection = "mercator",
  palette = "YlOrBr"
)
}
\arguments{
\item{plot_type}{"static" | "dynamic"}

\item{fill}{"avg_daily_new_cases_pre_vac" | "avg_daily_new_cases_post_vac" | "avg_daily_new_deaths_pre_vac" | "avg_daily_new_deaths_post_vac" - The map will be filled by this variable.}

\item{log_scale}{FALSE | TRUE - Apply a log transformation on the selected variable? Could be useful if variable contains outliers.}

\item{projection}{"mercator" | "globular" | "gilbert"}

\item{palette}{"YlOrBr" or any sequential palette from RColorBrewer: "YlOrRd", "YlOrBr", "YlGnBu", "YlGn", "Reds", "RdPu", "Purples", "PuRd", "PuBuGn", "PuBu", "OrRd", "Oranges", "Greys", "Greens", "GnBu", "BuPu", "BuGn", "Blues".}
}
\value{
Either a static map produced with ggplot2 or a dynamic one produced with leaflet.
}
\description{
For the countries that have vaccines, compare the average daily new cases (or deaths) of the days before the first vaccine dose administered and the days after that. Also, construct a 95\% confidence interval for each.
}
