% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query3}
\alias{query3}
\title{Query 3}
\usage{
query3(
  plot_type = "static",
  fill = "new_cases",
  log_scale = FALSE,
  projection = "mercator"
)
}
\arguments{
\item{plot_type}{"static" | "dynamic"}

\item{fill}{"new_cases" | "new_deaths" | "new_vaccinations" - The map will be filled by this variable.}

\item{log_scale}{FALSE | TRUE - Apply a log transformation on the selected variable? Could be useful if variable contains outliers.}

\item{projection}{"mercator" | "globular" | "gilbert"}
}
\value{
Either a static map produced with ggplot2 or a dynamic one produced with leaflet.
}
\description{
For each country, identify the top 3 peaks given a variable (e.g., daily new cases).
}
