% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query4}
\alias{query4}
\title{Query 4}
\usage{
query4(
  plot_type = "static",
  start = "outbreak_start",
  end = "most_recent",
  var = "new_cases",
  func = "sum",
  projection = "mercator"
)
}
\arguments{
\item{plot_type}{"static" | "dynamic"}

\item{start}{The start date of the time frame; this can be specified with one of the following:
\itemize{
\item "outbreak_start": The default.
\item "05-22-2020": Any date string in the format "\%m-\%d-\%Y".
\item c("new_cases" = 10000): The date when a variable first exceeded a threshold. This variable can be any of the numeric variables in the OWID COVID dataset found here: \url{https://covid.ourworldindata.org/data/owid-covid-data.csv}
}}

\item{end}{The end date of the time frame; this can be specified with one of the following:
\itemize{
\item "most_recent": The default.
\item "07-20-2021": Any date string in the format "\%m-\%d-\%Y".
\item c("new_deaths" = 1000): The date when a variable first exceeded a threshold.
}}

\item{var}{"new_cases" | "new_deaths" | "new_vaccinations" - or any other numeric variable in the OWID COVID dataset found here: \url{https://covid.ourworldindata.org/data/owid-covid-data.csv}}

\item{func}{"sum" | "mean" | "median" | "min" | "max"}

\item{projection}{"mercator" | "globular" | "gilbert"}
}
\value{
Either a static map produced with ggplot2 or a dynamic one produced with leaflet.
}
\description{
This is a more flexible query - essentially, a user can specify a time frame, then investigate any variable of their choice.
}
\examples{
query4(plot_type = "static", start = "outbreak_start", end = "most_recent", var = "new_cases", func = "sum", projection = "mercator")
query4(plot_type = "dynamic", start = c("new_deaths" = 1000), end = "04-33-2021", var = "new_vaccinations", func = "median", projection = "gilbert")
query4(plot_type = "dynamic", start = "03-03-2020", end = c("population" = 30000), var = "new_cases", func = "max", projection = "globular")
}
