% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query6}
\alias{query6}
\title{Query 6}
\usage{
query6(
  plot_type = "static",
  x = "new_cases",
  y = "new_deaths",
  country = "United States",
  start = NULL,
  end = NULL,
  use_bins = FALSE
)
}
\arguments{
\item{plot_type}{"static" | "dynamic"}

\item{x}{"new_cases" | "new_deaths" | "new_vaccinations" - or any other numeric variable in the OWID COVID dataset found here: \url{https://covid.ourworldindata.org/data/owid-covid-data.csv}}

\item{y}{Any numeric variable other than x.}

\item{country}{"United States" | "Canada" | "United Kingdom" - or any other country name.}

\item{start}{The start date of the time frame; by default (NULL) starts from the first available record in the OWID COVID dataset. Or specify any date string in the format "\%m-\%d-\%Y" such as "05-22-2020".}

\item{end}{The end date of the time frame; by default (NULL) includes the most recent record in the OWID COVID dataset. Or specify any date string in the format "\%m-\%d-\%Y" such as "07-22-2021".}

\item{use_bins}{TRUE | FALSE - apply binning on the points?}
}
\value{
Either a static plot produced with ggplot2 or a dynamic one produced with plotly.
}
\description{
Produces a scatterplot of any two numeric variables from the OWID COVID dataset. Can subset dataset by specifying one or more countries, a time frame; if dataset is large, apply binning.
}
