% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{query5}
\alias{query5}
\title{Query 5}
\usage{
query5(
  plot_type = "static",
  var = "new_cases",
  country = "United States",
  start = NULL,
  end = NULL
)
}
\arguments{
\item{plot_type}{"static" | "dynamic"}

\item{var}{"new_cases" | "new_deaths" | "new_vaccinations" - or any other numeric variable in the OWID COVID dataset found here: \url{https://covid.ourworldindata.org/data/owid-covid-data.csv}}

\item{country}{"United States" | "Canada" | "United Kingdom" - or any other country name.}

\item{start}{The start date of the time frame; by default (NULL) starts from the first available record in the OWID COVID dataset. Or specify any date string in the format "\%m-\%d-\%Y" such as "05-22-2020".}

\item{end}{The end date of the time frame; by default (NULL) includes the most recent record in the OWID COVID dataset. Or specify any date string in the format "\%m-\%d-\%Y" such as "07-22-2021".}
}
\value{
Either a static map produced with ggplot2 or a dynamic one produced with leaflet.
}
\description{
Produces a time-series plot of a variable for either one country or multiple countries. Time frame can be a custom argument.
}
